Server-side readables
For quests, we might need a lot of written notes and books used as hints (to
complement dialogs).  The current code requires a new sprite for each item,
even for items that look the same but have different readable content.

This change allows new item ids to be created in the server and use a
different item id to display in the client.  This will allow us to create
many books and notes in the server items using sprites in already deployed
clients.